Improve the workaround for pre-commit putting files at the end.
(of the command line, that is)

pre-commit always puts the files to check at the end of the hook’s
command line, which is incompatible with clang-tidy and oclint’s
unusual use of `--` to mean “the *list of files to check* ends here,
more options follow (from a different set of potential options than
the ones that go before the files to check).”

The hooks are already trying to work around this, but when pre-commit
passes more than one file to the hook, they only move one of the files.
Also, they move the file all the way to the beginning of the command
line, which seems to be unacceptable to newer versions of clang-tidy.

The improved workaround looks for a *second* `--` and moves everything
after that to right before the first `--`.  For instance

    -a -b -- -c -d -- foo.c bar.c

becomes

    -a -b foo.c bar.c -- -c -d

For compatibility’s sake, if there is only one `--`, then it scans for
the first subsequent argument that doesn’t begin with a dash and
guesses that this is the beginning of the file list:

    -a -b -- -c -d foo.c bar.c

also becomes

    -a -b foo.c bar.c -- -c -d

This guess may be wrong, for instance

    -a -b -- -I include -D USE_FSEEKO foo.c bar.c

will become

    -a -b include -D USE_FSEEKO foo.c bar.c -- -I

Therefore the hooks issue a warning message whenever only one `--` is found.